Added Map, Rooms, Changed Server-Client Protocol, Added Kill Count and Death Count on Players
Map object is located in the map package and contains an arraylist of
rooms.
The server-client protocol was changed so that the first element is the
number of players (For scalability) and the second is the current room.
Fun kill count and death count that can be used l8r